Migrate GenericData to rmgpy.tools.data submodule
