+++ rmgpy/tools/canteraModel.py
@@ -4,7 +4,8 @@ import numpy as np
 import cantera as ct
 from rmgpy.chemkin import getSpeciesIdentifier
 from rmgpy.species import Species
-from rmgpy.tools.plot import GenericData, GenericPlot, SimulationPlot
+from rmgpy.tools.data import GenericData
+from rmgpy.tools.plot import GenericPlot, SimulationPlot
 from rmgpy.quantity import Quantity
