+++ rmgpy/tools/plot.py
@@ -3,6 +3,7 @@ import matplotlib as mpl
 # This must be called before pylab, matplotlib.pyplot, or matplotlib.backends is imported
 mpl.use('Agg')
 import matplotlib.pyplot as plt
+from rmgpy.tools.data import GenericData
         
 def parseCSVData(csvFile):
     """
@@ -85,20 +86,7 @@ def findNearest(array, value):
     idx = (numpy.abs(array-value)).argmin()
     return idx
 
-class GenericData(object):
-    """
-    A generic data class for the purpose of plotting.
-    label is the original label for the data
-    data is the numpy array containing the data
-    The other parameters are additional common flags for data found in RMG
-    """
-    def __init__(self, label='', data=None, species=None, reaction=None, units=None, index=None):
-        self.label = label
-        self.data = data
-        self.species = species
-        self.reaction = reaction
-        self.units = units
-        self.index = index
+
 
 class GenericPlot(object):
     """
